Move /user/current to /auth/current endpoint
