+++ routes/usersRouter.js
@@ -1,26 +1,20 @@
-import { Router } from 'express';
+import { Router } from "express";
 
-import authenticateMiddleware from '../middlewares/authenticateMiddleware.js';
-import { uploadAvatarImageMiddleware } from '../middlewares/uploadImageMiddleware.js';
-import multerErrorHandlingMiddleware from '../middlewares/multerErrorHandlingMiddleware.js';
-import fileUploadValidationMiddleware from '../middlewares/fileUploadValidationMiddleware.js';
-import usersControllers from '../controllers/usersControllers.js';
+import authenticateMiddleware from "../middlewares/authenticateMiddleware.js";
+import { uploadAvatarImageMiddleware } from "../middlewares/uploadImageMiddleware.js";
+import multerErrorHandlingMiddleware from "../middlewares/multerErrorHandlingMiddleware.js";
+import fileUploadValidationMiddleware from "../middlewares/fileUploadValidationMiddleware.js";
+import usersControllers from "../controllers/usersControllers.js";
 
 const usersRouter = Router();
 
 usersRouter.get(
-  '/current',
-  authenticateMiddleware,
-  usersControllers.getCurrentUser
-);
-
-usersRouter.get(
-  '/:userId/followers',
+  "/:userId/followers",
   authenticateMiddleware,
   usersControllers.getFollowers
 );
 usersRouter.get(
-  '/:userId/following',
+  "/:userId/following",
   authenticateMiddleware,
   usersControllers.getFollowing
 );
@@ -32,9 +26,9 @@ usersRouter.get(
 );
 
 usersRouter.patch(
-  '/avatars',
+  "/avatars",
   authenticateMiddleware,
-  uploadAvatarImageMiddleware.single('avatar'),
+  uploadAvatarImageMiddleware.single("avatar"),
   multerErrorHandlingMiddleware,
   fileUploadValidationMiddleware,
   usersControllers.updateAvatar
